+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - On Windows, the desktop app now uses a custom title bar, and a custom menu bar powered by [OS-GUI.js](https://os-gui.js.org)
 - The background of the whole window is now light purple instead of white or black, matching the background color of main part of the app.
 
+### Added
+
+- Added a **Cursor Movement > Leash distance** setting. The cursor only moves when the head moves beyond the set radius (in pixels) from where the cursor last stopped. A circle visualization is shown in the screen overlay.
+
 ### Fixed
 
 - On Windows, the app should no longer freeze when using Tracky Mouse to click the Tracky Mouse app window's own minimize or close buttons, or right-clicking on its titlebar. ([issue #69](https://github.com/1j01/tracky-mouse/issues/69))